YoloPythonServer/requirements.txt
numpy
Pillow
opencv-python
opencv-contrib-python